Changed filename, updated multithreading
data_laod is now data_process
